+++ packages/create-content-sdk-app/src/templates/nextjs-app-router/src/components/content-sdk/CdpPageView.tsx
@@ -1,7 +1,7 @@
 'use client';
 import { CdpHelper, useSitecore } from '@sitecore-content-sdk/nextjs';
-import React, { useEffect } from 'react';
-import { pageView } from '@sitecore-content-sdk/events/browser';
+import { useEffect, JSX } from 'react';
+import { pageView } from '@sitecore-cloudsdk/events/browser';
 import config from 'sitecore.config';
 
 /**