+++ packages/create-content-sdk-app/src/templates/nextjs/src/components/content-sdk/CdpPageView.tsx
@@ -1,8 +1,7 @@
 ﻿import { CdpHelper, useSitecore } from '@sitecore-content-sdk/nextjs';
-import { useEffect } from 'react';
-import { pageView } from '@sitecore-content-sdk/events/browser';
+import { useEffect, JSX } from 'react';
+import { pageView } from '@sitecore-cloudsdk/events/browser';
 import config from 'sitecore.config';
-import { JSX } from 'react';
 
 /**
  * This is the CDP page view component.