+++ packages/nextjs/src/client/sitecore-nextjs-client.test.ts
@@ -322,4 +322,59 @@ describe('SitecoreClient', () => {
       ]);
     });
   });
+
+  describe('getPagePaths', () => {
+    it('should return static paths with site prefixes - multisite enabled by default', async () => {
+      const paths = [
+        { params: { path: ['_site_site-one', 'home'] }, locale: 'en' },
+        { params: { path: ['_site_site-one', 'about'] }, locale: 'en' },
+        { params: { path: ['_site_site-two', 'home'] }, locale: 'de-DE' },
+      ];
+
+      sitePathServiceStub.fetchSiteRoutes.resolves(paths);
+
+      const result = await sitecoreClient.getPagePaths(['site-one', 'site-two'], ['en', 'de-DE']);
+
+      expect(result).to.deep.equal(paths);
+    });
+
+    it('should return static paths with site prefixes when multisite enabled', async () => {
+      const paths = [
+        { params: { path: ['_site_site-one', 'home'] }, locale: 'en' },
+        { params: { path: ['_site_site-one', 'about'] }, locale: 'en' },
+        { params: { path: ['_site_site-two', 'home'] }, locale: 'de-DE' },
+      ];
+
+      sitePathServiceStub.fetchSiteRoutes.resolves(paths);
+
+      const result = await sitecoreClient.getPagePaths(
+        ['site-one', 'site-two'],
+        ['en', 'de-DE'],
+        undefined,
+        true
+      );
+
+      expect(result).to.deep.equal(paths);
+    });
+
+    it('should return static paths without site prefixes when multisite is disabled', async () => {
+      const paths = [
+        { params: { path: ['_site_site-one', 'home'] }, locale: 'en' },
+        { params: { path: ['_site_site-one', 'about'] }, locale: 'en' },
+        { params: { path: ['_site_site-two', 'home'] }, locale: 'de-DE' },
+      ];
+
+      const expectedPaths = [
+        { params: { path: ['home'] }, locale: 'en' },
+        { params: { path: ['about'] }, locale: 'en' },
+        { params: { path: ['home'] }, locale: 'de-DE' },
+      ];
+
+      sitePathServiceStub.fetchSiteRoutes.resolves(structuredClone(paths));
+
+      const result = await sitecoreClient.getPagePaths(['site-one'], ['en'], undefined, false);
+
+      expect(result).to.deep.equal(expectedPaths);
+    });
+  });
 });